requirements.txt
appdirs==1.4.3
argcomplete==1.10.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Pillow==6.1.0
pyserial==3.4
python-escpos==2.2.0
python-resize-image==1.1.19
pyusb==1.0.2
PyYAML==5.1.1
qrcode==6.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
